pyproject.toml
[tool.poetry]
name = "ml-repository-app"
version = "0.1.0"
description = "Model Repository Dash application"
authors = ["Carlos <csuarez@unicauca.edu.co>"]
packages = [
    { include = "model_registry/backend" },
    { include = "model_registry/api" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
dash = ">=2.11,<3.0"
plotly = ">=5.0.0,<6.0.0"
dash-bootstrap-components = "^1.6"
uvicorn = "^0.38.0"
fastapi = "^0.125.0"
gunicorn = "23.0.0"
pyyaml = "^6.0.3"
dotenv = "^0.9.9"
pandas = "^2.3.3"
tensorflow = "^2.20.0"
scikit-learn = "1.5.2"
dash-ag-grid = "^32.3.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.48"
pydantic-settings = "^2.13.1"
python-jose = "^3.5.0"
psycopg2 = "^2.9.11"
python-multipart = "^0.0.22"
pydantic = {extras = ["email"], version = "^2.12.5"}
alembic = "^1.18.4"
bcrypt = "4.0.1"
colorlog = "^6.10.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10"
isort = "^5.12"
flake8 = "7.3.0"
pytest = "^8.1"
pytest-cov = "^4.0"
ruff = "0.14.14"

[tool.poetry.scripts]
ml-repository-backend = "model_registry.backend.app_backend:main"
ml-repository-api = "model_registry.api.app_api:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88

# checkers configuration
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "B",  # bugbear
    "UP", # pyupgrade
]

ignore = [
    "E501",  # line too long (Black lo maneja)
]

exclude = [
    ".venv",
    "__pycache__",
    "build",
    "dist",
]
